Move Loader to be a separate component in a separate folder for reusability
